modify block number and timestamp type

+++ core/node/executor/primitives/src/lib.rs
@@ -48,7 +48,7 @@ where
 
 pub struct ContextEnv {
 	pub number: BlockNumber,
-	pub timestamp: u32,
+	pub timestamp: u64,
 }
 
 pub trait Context: Clone {
